fixing name changes, database connection

+++ CORE/Interfaces/IPlantRepository.cs
@@ -5,6 +5,6 @@ namespace CORE.Interfaces
     public interface IPlantRepository
     {
         Task<IEnumerable<Plant>> GetAllAsync();
-        Task<Plant?> GetByIdAsync(int id);
+        Task<Plant?> GetByIdAsync(long id);
     }
 }
