+++ CORE/Interfaces/IUserService.cs
@@ -7,10 +7,10 @@ namespace CORE.Interfaces
     public interface IUserService
     {
         Task<IEnumerable<UserDto>> GetAllUsersAsync();
-        Task<UserDto?> GetUserByIdAsync(int id);
+        Task<UserDto?> GetUserByIdAsync(long id);
         Task<UserDto> AddUserAsync(UserDto userDto);
         Task UpdateUserAsync(UserDto userDto);
-        Task DeleteUserAsync(int id);
+        Task DeleteUserAsync(long id);
     }
